整合SpringSecurity和JWT实现登录认证和授权

+++ src/main/java/com/yuan/mall/mbg/mapper/PmsBrandMapper.java
@@ -1,12 +1,14 @@
 package com.yuan.mall.mbg.mapper;
 
+
+import java.util.List;
+
 import com.yuan.mall.mbg.model.PmsBrand;
 import com.yuan.mall.mbg.model.PmsBrandExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PmsBrandMapper {
-    long countByExample(PmsBrandExample example);
+    int countByExample(PmsBrandExample example);
 
     int deleteByExample(PmsBrandExample example);
